Safe predictions from a generalized linear model

Safe predictions from a generalized linear model

# S3 method for glm
safe_predict(object, new_data, type = c("link", "conf_int",
  "response", "class", "prob"), se_fit = FALSE, level = 0.95,
  threshold = 0.5, ...)

Arguments

object
A glm object returned from a call to stats::glm().

new_data
A dataset as a data.frame or similar object. Must be coercable to a tibble using tibble::as_tibble().

type
What kind of predictions to return. Which predictions are available depends on the family of object.

"link" and "conf_int" are available for all families. "link" produces numeric predictions on the linear predictor scale. "conf_int" produces numeric predictions on the response scale and corresponding confidence bounds.

"response" results in a numeric prediction on the response scale and is available for families:

gaussian

Gamma

inverse.gaussian

poisson

quasipoisson

quasi

"class" results in hard class predictions and is only available for binomial and quasibinomial families

"prob" results in class predictions for each class and is only available for binomial and quasibinomial families

Default is "link".

se_fit
Logical indicating whether or not to also calculate standard errors for the fit at each point. These standard errors do not include the residual variance. Ignored when calculating a confidence or prediction interval.

level
A number between 0 and 1 to use as the confidence level when calculating confidence and prediction intervals. Ignored otherwise.

threshold
A number between 0 and 1 to use as a threshold for classification. When the class probability for the class corresponding to a positive event is greater than the threshold, the event will be classified as positive. Defaults to 0.5. See positive_class() for assistance determining which class is considered the positive class.

...
Unused.

Value

A tibble::tibble() with one row for each row of newdata. Predictions for observations with missing data will be NA. Returned tibble has different columns depending on type:

"response": .pred (numeric)

"class": .pred (factor)

"prob": .pred_{level} columns (numerics between 0 and 1)

"link": .pred (numeric)

"conf_int": .pred, .pred_lower, .pred_upper (all numeric)

"pred_int": .pred, .pred_lower, .pred_upper (all numeric)

For interval predictions, the tibble has additional attributes level and interval. The level is the same as the level argument and is between 0 and 1. interval is either "confidence" or "prediction". Some models may also set a method attribute to detail the method used to calculate the intervals.

Details

For GLMs, standard errors can only be calculated when type = "link".