Safe predictions from a smooth.spline object

We strongly recommend using mgcv::gam() instead of smooth.spline(). mgcv provides a feature-rich and frequently updated GAM implementation.

# S3 method for smooth.spline
safe_predict(object, new_data, type = "response",
  ...)

Arguments

object
A smooth.spline object returned from a call to stats::smooth.spline().

new_data
A numeric vector. Can contain NAs, although the presence of any NAs may slow down prediction.

type
What kind of predictions to return. Options are:

"response" (default): Standard predictions from smoothing splines.

Value

A tibble::tibble() with one row for each row of new_data. Predictions for observations with missing data will be NA. Returned tibble has different columns depending on type:

"response":

univariate outcome: .pred (numeric)

multivariate outcomes: .pred_{outcome name} (numeric) for each outcome

"class": .pred_class (factor)

"prob": .pred_{level} columns (numerics between 0 and 1)

"link": .pred (numeric)

"conf_int": .pred, .pred_lower, .pred_upper (all numeric)

"pred_int": .pred, .pred_lower, .pred_upper (all numeric)

If you request standard errors with std_error = TRUE, an additional column .std_error.

For interval predictions, the tibble has additional attributes level and interval. The level is the same as the level argument and is between 0 and 1. interval is either "confidence" or "prediction". Some models may also set a method attribute to detail the method used to calculate the intervals.

Uncertainty in predictions

If you use mgcv, mgcv::predict.gam() has a se.fit argument that will allow you get standard errors for predictions.

If you insist on using smooth.spline(), your best bet to estimate uncertainty in predictions is bootstrapping. See

vignette("bootstrapping", package = "safepredict")

for worked examples and details on how you might do this.

Derivatives of smooths

While stats::::predict.smooth.spline() has a deriv argument we do not support it. If you would like to estimate derivates of smooths we recommend using the gratia package, and in particular gratia::fderiv(). At the time of writing, gratia is not yet on CRAN, but can be installed from Github with devtools::install_github(gavinsimpson/gratia). Additional documentation on gratia is available here.

Examples

fit <- smooth.spline(mtcars$mpg, mtcars$wt, cv = TRUE)

#> Warning: cross-validation with non-unique 'x' values seems doubtful

safe_predict(fit, c(30, NA, 40))

#> # A tibble: 3 x 1
#>   .pred
#>   <dbl>
#> 1  1.91
#> 2 NA   
#> 3  1.06

# the following will fail, however

# NOT RUN {
predict(fit, c(30, NA, 40))
# }

# however, we recommend using mgcv instead
library(mgcv)

#> Loading required package: nlme

#> This is mgcv 1.8-23. For overview type 'help("mgcv-package")'.

fit2 <- gam(mpg ~ s(wt), data = mtcars)
predict(fit2, mtcars)  # TODO: update with safe_predict once implemented

#>           Mazda RX4       Mazda RX4 Wag          Datsun 710      Hornet 4 Drive 
#>            23.03365            21.30454            25.24007            19.21380 
#>   Hornet Sportabout             Valiant          Duster 360           Merc 240D 
#>            17.95427            17.84824            17.28478            19.36025 
#>            Merc 230            Merc 280           Merc 280C          Merc 450SE 
#>            19.59685            17.95427            17.95427            15.11330 
#>          Merc 450SL         Merc 450SLC  Cadillac Fleetwood Lincoln Continental 
#>            16.52446            16.30057            11.46772            11.00221 
#>   Chrysler Imperial            Fiat 128         Honda Civic      Toyota Corolla 
#>            11.21332            26.15766            30.67862            28.98223 
#>       Toyota Corona    Dodge Challenger         AMC Javelin          Camaro Z28 
#>            24.15407            17.53673            17.98094            16.03997 
#>    Pontiac Firebird           Fiat X1-9       Porsche 914-2        Lotus Europa 
#>            16.01864            28.20817            26.62028            31.46338 
#>      Ford Pantera L        Ferrari Dino       Maserati Bora          Volvo 142E 
#>            19.47820            21.99848            17.28478            21.93130 

# to get estimated derivatives from smooths use gratia,
# which extends mgcv

library(gratia)  # TODO: gratia isn't on CRAN yet
fderiv(fit2)

#> $derivatives
#> $derivatives$`s(wt)`
#> $derivatives$`s(wt)`$deriv
#>          [,1]
#> 1   -7.692424
#> 2   -7.692577
#> 3   -7.693035
#> 4   -7.693799
#> 5   -7.694869
#> 6   -7.696244
#> 7   -7.697910
#> 8   -7.699837
#> 9   -7.702024
#> 10  -7.704471
#> 11  -7.707178
#> 12  -7.710145
#> 13  -7.713372
#> 14  -7.716860
#> 15  -7.720607
#> 16  -7.724614
#> 17  -7.728882
#> 18  -7.733298
#> 19  -7.737429
#> 20  -7.741233
#> 21  -7.744708
#> 22  -7.747856
#> 23  -7.750582
#> 24  -7.752381
#> 25  -7.753178
#> 26  -7.752972
#> 27  -7.751765
#> 28  -7.749556
#> 29  -7.746344
#> 30  -7.742131
#> 31  -7.736915
#> 32  -7.730697
#> 33  -7.723476
#> 34  -7.714985
#> 35  -7.705042
#> 36  -7.693647
#> 37  -7.680644
#> 38  -7.665900
#> 39  -7.649416
#> 40  -7.631191
#> 41  -7.611227
#> 42  -7.589522
#> 43  -7.566126
#> 44  -7.541102
#> 45  -7.514449
#> 46  -7.486168
#> 47  -7.456259
#> 48  -7.424721
#> 49  -7.391555
#> 50  -7.356847
#> 51  -7.321011
#> 52  -7.284100
#> 53  -7.246114
#> 54  -7.207052
#> 55  -7.166916
#> 56  -7.125706
#> 57  -7.083420
#> 58  -7.040222
#> 59  -6.996630
#> 60  -6.952682
#> 61  -6.908379
#> 62  -6.863719
#> 63  -6.818704
#> 64  -6.773333
#> 65  -6.727607
#> 66  -6.681837
#> 67  -6.636564
#> 68  -6.591834
#> 69  -6.547648
#> 70  -6.504006
#> 71  -6.460932
#> 72  -6.418501
#> 73  -6.376716
#> 74  -6.335579
#> 75  -6.295088
#> 76  -6.255245
#> 77  -6.216048
#> 78  -6.177498
#> 79  -6.139595
#> 80  -6.102339
#> 81  -6.065730
#> 82  -6.029768
#> 83  -5.994453
#> 84  -5.959785
#> 85  -5.925622
#> 86  -5.891433
#> 87  -5.856678
#> 88  -5.820912
#> 89  -5.783755
#> 90  -5.745120
#> 91  -5.705008
#> 92  -5.663417
#> 93  -5.620348
#> 94  -5.575802
#> 95  -5.529778
#> 96  -5.482275
#> 97  -5.433295
#> 98  -5.382837
#> 99  -5.330908
#> 100 -5.277862
#> 101 -5.224143
#> 102 -5.169968
#> 103 -5.115337
#> 104 -5.060440
#> 105 -5.005575
#> 106 -4.950773
#> 107 -4.896406
#> 108 -4.842588
#> 109 -4.789318
#> 110 -4.736597
#> 111 -4.684424
#> 112 -4.632800
#> 113 -4.581724
#> 114 -4.531202
#> 115 -4.481412
#> 116 -4.432443
#> 117 -4.384325
#> 118 -4.337165
#> 119 -4.290972
#> 120 -4.245743
#> 121 -4.201462
#> 122 -4.158123
#> 123 -4.115728
#> 124 -4.074276
#> 125 -4.033766
#> 126 -3.994200
#> 127 -3.955576
#> 128 -3.917896
#> 129 -3.881158
#> 130 -3.845364
#> 131 -3.810512
#> 132 -3.776434
#> 133 -3.742877
#> 134 -3.709839
#> 135 -3.677319
#> 136 -3.645318
#> 137 -3.613837
#> 138 -3.582873
#> 139 -3.552429
#> 140 -3.522504
#> 141 -3.493097
#> 142 -3.464209
#> 143 -3.435840
#> 144 -3.407990
#> 145 -3.380658
#> 146 -3.353845
#> 147 -3.327551
#> 148 -3.301776
#> 149 -3.276520
#> 150 -3.251782
#> 151 -3.227564
#> 152 -3.203864
#> 153 -3.180683
#> 154 -3.158020
#> 155 -3.135877
#> 156 -3.114252
#> 157 -3.093146
#> 158 -3.072559
#> 159 -3.052490
#> 160 -3.032941
#> 161 -3.013910
#> 162 -2.995398
#> 163 -2.977405
#> 164 -2.959931
#> 165 -2.942975
#> 166 -2.926538
#> 167 -2.910620
#> 168 -2.895221
#> 169 -2.880341
#> 170 -2.865979
#> 171 -2.852136
#> 172 -2.838812
#> 173 -2.826007
#> 174 -2.813721
#> 175 -2.801953
#> 176 -2.790704
#> 177 -2.779974
#> 178 -2.769763
#> 179 -2.760071
#> 180 -2.750897
#> 181 -2.742242
#> 182 -2.734106
#> 183 -2.726489
#> 184 -2.719390
#> 185 -2.712811
#> 186 -2.706750
#> 187 -2.701208
#> 188 -2.696185
#> 189 -2.691680
#> 190 -2.687694
#> 191 -2.684227
#> 192 -2.681248
#> 193 -2.678701
#> 194 -2.676585
#> 195 -2.674902
#> 196 -2.673650
#> 197 -2.672741
#> 198 -2.672092
#> 199 -2.671702
#> 200 -2.671572
#> 
#> $derivatives$`s(wt)`$se.deriv
#>         1         2         3         4         5         6         7         8 
#> 1.9410359 1.9397027 1.9357157 1.9291126 1.9199571 1.9083404 1.8942045 1.8773359 
#>         9        10        11        12        13        14        15        16 
#> 1.8578899 1.8360650 1.8120988 1.7862724 1.7589170 1.7304210 1.7012369 1.6718899 
#>        17        18        19        20        21        22        23        24 
#> 1.6429851 1.6152701 1.5893914 1.5654084 1.5433524 1.5232678 1.5051278 1.4882863 
#>        25        26        27        28        29        30        31        32 
#> 1.4721339 1.4561800 1.4400186 1.4233324 1.4058976 1.3875885 1.3683833 1.3483703 
#>        33        34        35        36        37        38        39        40 
#> 1.3277583 1.3069433 1.2861866 1.2657179 1.2460259 1.2273364 1.2096593 1.1930640 
#>        41        42        43        44        45        46        47        48 
#> 1.1776834 1.1637163 1.1513982 1.1405816 1.1309537 1.1222599 1.1143085 1.1069754 
#>        49        50        51        52        53        54        55        56 
#> 1.1002070 1.0939566 1.0877898 1.0813372 1.0743569 1.0667127 1.0583776 1.0494371 
#>        57        58        59        60        61        62        63        64 
#> 1.0400944 1.0306716 1.0213729 1.0122687 1.0035263 0.9954216 0.9883422 0.9827891 
#>        65        66        67        68        69        70        71        72 
#> 0.9793730 0.9786159 0.9802074 0.9835393 0.9880974 0.9934724 0.9992742 1.0048092 
#>        73        74        75        76        77        78        79        80 
#> 1.0095064 1.0129595 1.0149115 1.0152489 1.0139985 1.0113277 1.0075476 1.0031189 
#>        81        82        83        84        85        86        87        88 
#> 0.9986582 0.9949431 0.9929124 0.9936572 0.9982579 1.0069973 1.0191458 1.0332622 
#>        89        90        91        92        93        94        95        96 
#> 1.0475564 1.0606535 1.0716436 1.0799474 1.0852875 1.0876697 1.0873764 1.0849708 
#>        97        98        99       100       101       102       103       104 
#> 1.0813086 1.0775537 1.0751923 1.0758202 1.0797801 1.0864551 1.0951805 1.1052341 
#>       105       106       107       108       109       110       111       112 
#> 1.1155586 1.1253872 1.1338883 1.1403634 1.1444356 1.1460077 1.1452557 1.1426300 
#>       113       114       115       116       117       118       119       120 
#> 1.1388642 1.1349938 1.1323291 1.1316175 1.1334488 1.1378524 1.1445341 1.1531706 
#>       121       122       123       124       125       126       127       128 
#> 1.1624970 1.1710064 1.1775067 1.1810691 1.1810026 1.1768317 1.1682813 1.1552705 
#>       129       130       131       132       133       134       135       136 
#> 1.1379165 1.1165510 1.0917495 1.0666207 1.0448819 1.0264833 1.0113136 0.9992346 
#>       137       138       139       140       141       142       143       144 
#> 0.9900833 0.9836757 0.9798108 0.9782756 0.9788500 0.9813120 0.9854421 0.9910270 
#>       145       146       147       148       149       150       151       152 
#> 0.9978629 1.0057585 1.0145357 1.0240319 1.0341000 1.0446086 1.0554422 1.0665007 
#>       153       154       155       156       157       158       159       160 
#> 1.0776988 1.0889654 1.1002432 1.1114876 1.1226665 1.1337593 1.1447567 1.1556598 
#>       161       162       163       164       165       166       167       168 
#> 1.1664797 1.1772372 1.1879623 1.1986936 1.2094784 1.2203716 1.2314360 1.2427415 
#>       169       170       171       172       173       174       175       176 
#> 1.2543649 1.2663889 1.2789026 1.2919997 1.3057789 1.3203426 1.3357962 1.3522475 
#>       177       178       179       180       181       182       183       184 
#> 1.3698057 1.3885804 1.4086809 1.4302151 1.4532885 1.4780034 1.5044582 1.5327463 
#>       185       186       187       188       189       190       191       192 
#> 1.5629559 1.5951688 1.6294609 1.6659008 1.7045507 1.7454655 1.7886932 1.8322082 
#>       193       194       195       196       197       198       199       200 
#> 1.8721078 1.9077743 1.9387643 1.9647283 1.9852858 2.0001200 2.0090795 2.0120756 
#> 
#> $derivatives$`s(wt)`$Xi
#>     (Intercept)       s(wt).1     s(wt).2     s(wt).3      s(wt).4
#> 1             0  0.1504135771 -0.50111896 -0.23128146  0.550338966
#> 2             0  0.1501863123 -0.50117781 -0.22927622  0.551071417
#> 3             0  0.1495046442 -0.50135407 -0.22326052  0.553268134
#> 4             0  0.1483684930 -0.50164777 -0.21323500  0.556930611
#> 5             0  0.1467779054 -0.50205899 -0.19919919  0.562057136
#> 6             0  0.1447328435 -0.50258786 -0.18115298  0.568649077
#> 7             0  0.1421084961 -0.50341521 -0.15862222  0.576646513
#> 8             0  0.1386561976 -0.50490141 -0.13066099  0.585932103
#> 9             0  0.1343690026 -0.50705687 -0.09724347  0.596502507
#> 10            0  0.1292469221 -0.50988151 -0.05836952  0.608357932
#> 11            0  0.1232900182 -0.51337502 -0.01403923  0.621497941
#> 12            0  0.1164982133 -0.51753791  0.03574763  0.635922962
#> 13            0  0.1088715540 -0.52236971  0.09099077  0.651632913
#> 14            0  0.1004100514 -0.52787054  0.15169027  0.668627483
#> 15            0  0.0911136522 -0.53404047  0.21784607  0.686907006
#> 16            0  0.0809823986 -0.54087956  0.28945842  0.706471421
#> 17            0  0.0700162928 -0.54838772  0.36652716  0.727320452
#> 18            0  0.0581823545 -0.55668875  0.44851891  0.749063241
#> 19            0  0.0453526794 -0.56626355  0.53336323  0.770178690
#> 20            0  0.0315145043 -0.57716019  0.62085255  0.790514947
#> 21            0  0.0166677627 -0.58937879  0.71098696  0.810072334
#> 22            0  0.0008125101 -0.60291909  0.80376657  0.828850250
#> 23            0 -0.0160592961 -0.61782235  0.89866214  0.846565554
#> 24            0 -0.0339907569 -0.63430941  0.99282737  0.861691888
#> 25            0 -0.0529882294 -0.65241280  1.08583869  0.874001813
#> 26            0 -0.0730517691 -0.67213285  1.17769592  0.883495498
#> 27            0 -0.0941813605 -0.69346945  1.26839927  0.890172772
#> 28            0 -0.1163769880 -0.71642248  1.35794864  0.894034097
#> 29            0 -0.1396386651 -0.74099217  1.44634391  0.895079048
#> 30            0 -0.1639663805 -0.76717826  1.53358520  0.893307718
#> 31            0 -0.1893601542 -0.79498098  1.61967258  0.888720199
#> 32            0 -0.2158199708 -0.82440021  1.70460595  0.881316382
#> 33            0 -0.2433449309 -0.85543184  1.78836860  0.871093512
#> 34            0 -0.2718224135 -0.88755672  1.86886920  0.857705633
#> 35            0 -0.3011742544 -0.92041472  1.94465628  0.840912107
#> 36            0 -0.3314001917 -0.95400457  2.01572691  0.820713192
#> 37            0 -0.3623617983 -0.98767371  2.08048173  0.797117974
#> 38            0 -0.3939423787 -1.02087187  2.13757243  0.770133914
#> 39            0 -0.4261419129 -1.05359904  2.18699886  0.739761258
#> 40            0 -0.4589604186 -1.08585526  2.22876121  0.705999867
#> 41            0 -0.4923978847 -1.11764059  2.26285942  0.668849830
#> 42            0 -0.5264543113 -1.14895491  2.28929342  0.628311145
#> 43            0 -0.5609752685 -1.17909769  2.30754008  0.584920485
#> 44            0 -0.5957648797 -1.20718043  2.31693591  0.539358516
#> 45            0 -0.6308224876 -1.23320008  2.31747870  0.491627753
#> 46            0 -0.6661480922 -1.25715673  2.30916841  0.441727985
#> 47            0 -0.7017416936 -1.27905037  2.29200512  0.389659219
#> 48            0 -0.7376032807 -1.29888094  2.26598868  0.335421588
#> 49            0 -0.7737328600 -1.31664838  2.23111917  0.279015135
#> 50            0 -0.8100696380 -1.33213935  2.18760816  0.220802999
#> 51            0 -0.8463244106 -1.34433842  2.13646198  0.162513817
#> 52            0 -0.8824597231 -1.35311406  2.07781099  0.104371364
#> 53            0 -0.9184755601 -1.35846616  2.01165511  0.046375813
#> 54            0 -0.9543719326 -1.36039483  1.93799448 -0.011472953
#> 55            0 -0.9901488451 -1.35890014  1.85682902 -0.069175023
#> 56            0 -1.0258062755 -1.35398189  1.76815873 -0.126730145
#> 57            0 -1.0613442503 -1.34564022  1.67198365 -0.184138592
#> 58            0 -1.0966879826 -1.33380975  1.56911955 -0.240810781
#> 59            0 -1.1316001103 -1.31828298  1.46215595 -0.294875711
#> 60            0 -1.1660630794 -1.29904465  1.35128467 -0.346195011
#> 61            0 -1.2000768701 -1.27609460  1.23650558 -0.394768535
#> 62            0 -1.2336414912 -1.24943299  1.11781884 -0.440596358
#> 63            0 -1.2667569393 -1.21905973  0.99522430 -0.483678411
#> 64            0 -1.2994232157 -1.18497484  0.86872207 -0.524014803
#> 65            0 -1.3316401315 -1.14717908  0.73831584 -0.561604013
#> 66            0 -1.3632534446 -1.10631008  0.60703157 -0.595345567
#> 67            0 -1.3939926946 -1.06353010  0.48023548 -0.623342163
#> 68            0 -1.4238336987 -1.01894694  0.35841295 -0.645426756
#> 69            0 -1.4527764447 -0.97256066  0.24156403 -0.661599339
#> 70            0 -1.4808209370 -0.92437114  0.12968862 -0.671859840
#> 71            0 -1.5079257565 -0.87471214  0.02376055 -0.676154793
#> 72            0 -1.5339702586 -0.82455550 -0.07338419 -0.674328045
#> 73            0 -1.5589467384 -0.77396333 -0.16156428 -0.666369482
#> 74            0 -1.5828551914 -0.72293564 -0.24077967 -0.652279234
#> 75            0 -1.6056956254 -0.67147248 -0.31103035 -0.632057390
#> 76            0 -1.6274680381 -0.61957379 -0.37231639 -0.605703802
#> 77            0 -1.6481724291 -0.56723959 -0.42463770 -0.573218476
#> 78            0 -1.6678087889 -0.51446988 -0.46799438 -0.534601473
#> 79            0 -1.6863771307 -0.46126465 -0.50238633 -0.489852700
#> 80            0 -1.7038774475 -0.40762395 -0.52781358 -0.438972343
#> 81            0 -1.7203097430 -0.35354771 -0.54427620 -0.381960241
#> 82            0 -1.7356740134 -0.29903597 -0.55177411 -0.318816402
#> 83            0 -1.7499702590 -0.24408870 -0.55030735 -0.249540870
#> 84            0 -1.7631984786 -0.18870595 -0.53987588 -0.174133652
#> 85            0 -1.7753314960 -0.13303101 -0.52045223 -0.093445547
#> 86            0 -1.7862669732 -0.07757972 -0.49200858 -0.010717356
#> 87            0 -1.7959011946 -0.02278225 -0.45481131  0.070627542
#> 88            0 -1.8041469890  0.03107609 -0.40932473  0.147619456
#> 89            0 -1.8109261092  0.08382914 -0.35625203  0.217507154
#> 90            0 -1.8162208002  0.13544083 -0.29575816  0.279664322
#> 91            0 -1.8200310512  0.18591112 -0.22784302  0.334091037
#> 92            0 -1.8223568758  0.23524004 -0.15250665  0.380787266
#> 93            0 -1.8231982656  0.28342759 -0.06974906  0.419753026
#> 94            0 -1.8225552167  0.33047379  0.02042973  0.450988356
#> 95            0 -1.8204277344  0.37637861  0.11802975  0.474493208
#> 96            0 -1.8168158206  0.42114205  0.22305100  0.490267558
#> 97            0 -1.8117194728  0.46476410  0.33549349  0.498311462
#> 98            0 -1.8051386985  0.50724480  0.45535718  0.498624938
#> 99            0 -1.7970739424  0.54859002  0.58257846  0.491198701
#> 100           0 -1.7875517561  0.58910619  0.71376254  0.475592333
#> 101           0 -1.7766151700  0.62910380  0.84493775  0.451539389
#> 102           0 -1.7642918981  0.66868226  0.97436825  0.419108575
#> 103           0 -1.7505820610  0.70784198  1.10204705  0.378300123
#> 104           0 -1.7355345305  0.74649825  1.22683548  0.329702629
#> 105           0 -1.7192258062  0.78451838  1.34694955  0.274238017
#> 106           0 -1.7016678211  0.82186598  1.46222096  0.212064007
#> 107           0 -1.6830021088  0.85809313  1.57076013  0.145065182
#> 108           0 -1.6632718391  0.89306307  1.67198983  0.073817168
#> 109           0 -1.6424770632  0.92677578  1.76591016 -0.001680032
#> 110           0 -1.6206177156  0.95923128  1.85252101 -0.081426418
#> 111           0 -1.5976938506  0.99042959  1.93182242 -0.165421948
#> 112           0 -1.5737054582  1.02037063  2.00381447 -0.253666722
#> 113           0 -1.5486525018  1.04905450  2.06849703 -0.346160663
#> 114           0 -1.5225418104  1.07645725  2.12586003 -0.442826298
#> 115           0 -1.4956084926  1.10174868  2.17555074 -0.540968328
#> 116           0 -1.4679687910  1.12451837  2.21739476 -0.639254479
#> 117           0 -1.4396997827  1.14456256  2.25145745 -0.736929015
#> 118           0 -1.4110665092  1.16118017  2.27796346 -0.831391486
#> 119           0 -1.3820912459  1.17431227  2.29693166 -0.922423182
#> 120           0 -1.3528532450  1.18391927  2.30869187 -1.009512557
#> 121           0 -1.3238404617  1.18978752  2.31531399 -1.089551631
#> 122           0 -1.2951514927  1.19187946  2.31722354 -1.161921042
#> 123           0 -1.2667863691  1.19019506  2.31442069 -1.226620833
#> 124           0 -1.2387450621  1.18473432  2.30690532 -1.283650941
#> 125           0 -1.2110275716  1.17549728  2.29467720 -1.333011441
#> 126           0 -1.1836338842  1.16248390  2.27773679 -1.374702231
#> 127           0 -1.1565640312  1.14569419  2.25608379 -1.408723413
#> 128           0 -1.1298180125  1.12512818  2.22971817 -1.435074969
#> 129           0 -1.1033958169  1.10078584  2.19864015 -1.453756812
#> 130           0 -1.0772974313  1.07266716  2.16284957 -1.464769039
#> 131           0 -1.0515228821  1.04077216  2.12234641 -1.468111602
#> 132           0 -1.0260875527  1.00693316  2.07919098 -1.466691050
#> 133           0 -1.0010143381  0.97387340  2.03644522 -1.464827007
#> 134           0 -0.9763034803  0.94161835  1.99413772 -1.462559990
#> 135           0 -0.9519549793  0.91016806  1.95226850 -1.459890004
#> 136           0 -0.9279687885  0.87952253  1.91083766 -1.456817105
#> 137           0 -0.9043449634  0.84968174  1.86984514 -1.453341235
#> 138           0 -0.8810834684  0.82064572  1.82929084 -1.449462357
#> 139           0 -0.8581843258  0.79241441  1.78917490 -1.445180562
#> 140           0 -0.8356475201  0.76498788  1.74949737 -1.440495770
#> 141           0 -0.8134730467  0.73836608  1.71025804 -1.435408018
#> 142           0 -0.7916609412  0.71254901  1.67145708 -1.429917287
#> 143           0 -0.7702111660  0.68753674  1.63309449 -1.424023656
#> 144           0 -0.7491237142  0.66332917  1.59517006 -1.417727061
#> 145           0 -0.7283986370  0.63992641  1.55768402 -1.411027459
#> 146           0 -0.7080359010  0.61732832  1.52063645 -1.403924879
#> 147           0 -0.6880354819  0.59553501  1.48402694 -1.396419425
#> 148           0 -0.6683974330  0.57454645  1.44785605 -1.388510962
#> 149           0 -0.6491216897  0.55436268  1.41212321 -1.380199535
#> 150           0 -0.6302082967  0.53498361  1.37682873 -1.371485134
#> 151           0 -0.6116572915  0.51640933  1.34197269 -1.362367773
#> 152           0 -0.5934685854  0.49863978  1.30755489 -1.352847431
#> 153           0 -0.5756422317  0.48167495  1.27357546 -1.342924140
#> 154           0 -0.5581782059  0.46551489  1.24003417 -1.332597938
#> 155           0 -0.5410765680  0.45015957  1.20693132 -1.321868731
#> 156           0 -0.5243372292  0.43560900  1.17426687 -1.310736600
#> 157           0 -0.5079602561  0.42186320  1.14204061 -1.299201410
#> 158           0 -0.4919456220  0.40892213  1.11025281 -1.287263345
#> 159           0 -0.4762933181  0.39678579  1.07890312 -1.274922261
#> 160           0 -0.4610033488  0.38545424  1.04799196 -1.262178264
#> 161           0 -0.4460757586  0.37492740  1.01751904 -1.249031248
#> 162           0 -0.4315104851  0.36520538  0.98748433 -1.235481376
#> 163           0 -0.4173075641  0.35628802  0.95788804 -1.221528525
#> 164           0 -0.4034669820  0.34817546  0.92873006 -1.207172615
#> 165           0 -0.3899887169  0.34086761  0.90001037 -1.192413710
#> 166           0 -0.3768728374  0.33436453  0.87172906 -1.177251994
#> 167           0 -0.3641192881  0.32866619  0.84388618 -1.161687240
#> 168           0 -0.3517280933  0.32377264  0.81648135 -1.145719466
#> 169           0 -0.3396991999  0.31968375  0.78951493 -1.129348837
#> 170           0 -0.3280326721  0.31639969  0.76298693 -1.112575219
#> 171           0 -0.3167284879  0.31392035  0.73689693 -1.095398632
#> 172           0 -0.3057866405  0.31224577  0.71124560 -1.077819074
#> 173           0 -0.2952071387  0.31137596  0.68603240 -1.059836618
#> 174           0 -0.2849899805  0.31131082  0.66125759 -1.041451032
#> 175           0 -0.2751351680  0.31205053  0.63692110 -1.022662675
#> 176           0 -0.2656426790  0.31359493  0.61302282 -1.003471215
#> 177           0 -0.2565125268  0.31594412  0.58956301 -0.983876833
#> 178           0 -0.2477448047  0.31909803  0.56654159 -0.963879547
#> 179           0 -0.2393393350  0.32305670  0.54395826 -0.943479244
#> 180           0 -0.2312962155  0.32782005  0.52181334 -0.922676029
#> 181           0 -0.2236154528  0.33338825  0.50010672 -0.901469719
#> 182           0 -0.2162970136  0.33976109  0.47883847 -0.879860649
#> 183           0 -0.2093409845  0.34693882  0.45800866 -0.857848439
#> 184           0 -0.2027472257  0.35492119  0.43761675 -0.835433382
#> 185           0 -0.1965158636  0.36370828  0.41766341 -0.812615342
#> 186           0 -0.1906468272  0.37330022  0.39814845 -0.789394237
#> 187           0 -0.1851401321  0.38369684  0.37907177 -0.765770352
#> 188           0 -0.1799957627  0.39489830  0.36043325 -0.741743371
#> 189           0 -0.1752137080  0.40690437  0.34223340 -0.717313531
#> 190           0 -0.1707940456  0.41971520  0.32447170 -0.692480531
#> 191           0 -0.1667367222  0.43333091  0.30714810 -0.667244753
#> 192           0 -0.1630868263  0.44693033  0.29083971 -0.642796207
#> 193           0 -0.1599218935  0.45910435  0.27653539 -0.621178202
#> 194           0 -0.1572431851  0.46982895  0.26425226 -0.602425558
#> 195           0 -0.1550507522  0.47910415  0.25399058 -0.586538127
#> 196           0 -0.1533444616  0.48692981  0.24574966 -0.573516157
#> 197           0 -0.1520622162  0.49313758  0.23941967 -0.563376015
#> 198           0 -0.1511462711  0.49757167  0.23489854 -0.556133135
#> 199           0 -0.1505967351  0.50023217  0.23218548 -0.551787165
#> 200           0 -0.1504135749  0.50111895  0.23128149 -0.550338664
#>           s(wt).5     s(wt).6      s(wt).7     s(wt).8 s(wt).9
#> 1   -0.4292033606 -0.48271276 -0.417453504  1.03173203 1.03837
#> 2   -0.4268952924 -0.48877903 -0.412763383  1.03231923 1.03837
#> 3   -0.4199711579 -0.50697972 -0.398698531  1.03408148 1.03837
#> 4   -0.4084329486 -0.53731342 -0.375256671  1.03701878 1.03837
#> 5   -0.3922779046 -0.57978142 -0.342438442  1.04113087 1.03837
#> 6   -0.3715076391 -0.63438210 -0.300240477  1.04641760 1.03837
#> 7   -0.3459934261 -0.70001356 -0.249624135  1.05275654 1.03837
#> 8   -0.3154813999 -0.77447888 -0.192503866  1.05990397 1.03837
#> 9   -0.2799637699 -0.85771622 -0.128923314  1.06785221 1.03837
#> 10  -0.2394407328 -0.94972485 -0.058886949  1.07660152 1.03837
#> 11  -0.1939121097 -1.05050702  0.017601687  1.08615244 1.03837
#> 12  -0.1433777752 -1.16006056  0.100550863  1.09650407 1.03837
#> 13  -0.0878381756 -1.27838699  0.189953695  1.10765719 1.03837
#> 14  -0.0272932388 -1.40548571  0.285810693  1.11961161 1.03837
#> 15   0.0382570053 -1.54135680  0.388124393  1.13236722 1.03837
#> 16   0.1088127910 -1.68599991  0.496895034  1.14592388 1.03837
#> 17   0.1843743413 -1.83941560  0.612120127  1.16028184 1.03837
#> 18   0.2638359609 -1.99807369  0.730956538  1.17510029 1.03837
#> 19   0.3429095818 -2.14826823  0.842351874  1.18905610 1.03837
#> 20   0.4211649629 -2.28862723  0.945199903  1.20201699 1.03837
#> 21   0.4986020374 -2.41915011  1.039502846  1.21398273 1.03837
#> 22   0.5752211973 -2.53983809  1.125257381  1.22495366 1.03837
#> 23   0.6502119598 -2.64872579  1.201182374  1.23471783 1.03837
#> 24   0.7192160290 -2.73524495  1.260375286  1.24213602 1.03837
#> 25   0.7815842018 -2.79782372  1.301802632  1.24703929 1.03837
#> 26   0.8373161520 -2.83646074  1.325469940  1.24942710 1.03837
#> 27   0.8864130718 -2.85115659  1.331374133  1.24929960 1.03837
#> 28   0.9288737490 -2.84191145  1.319515548  1.24665701 1.03837
#> 29   0.9646987653 -2.80872513  1.289897203  1.24149893 1.03837
#> 30   0.9938878520 -2.75159777  1.242513833  1.23382583 1.03837
#> 31   1.0164414366 -2.67052920  1.177370150  1.22363729 1.03837
#> 32   1.0323591582 -2.56551929  1.094463651  1.21093351 1.03837
#> 33   1.0416381246 -2.43660445  0.993847797  1.19571561 1.03837
#> 34   1.0438882070 -2.28833816  0.882183377  1.17812947 1.03837
#> 35   1.0388400917 -2.12387930  0.764092954  1.15827588 1.03837
#> 36   1.0264942807 -1.94324101  0.639589333  1.13615590 1.03837
#> 37   1.0075667700 -1.75339238  0.516074998  1.11221784 1.03837
#> 38   0.9826608749 -1.56020979  0.399788593  1.08683997 1.03837
#> 39   0.9517764377 -1.36369286  0.290730671  1.06002223 1.03837
#> 40   0.9149135916 -1.16384158  0.188900946  1.03176460 1.03837
#> 41   0.8720728562 -0.96065585  0.094300795  1.00206694 1.03837
#> 42   0.8232533222 -0.75413599  0.006928582  0.97092950 1.03837
#> 43   0.7704694371 -0.55153565 -0.069199251  0.93908753 1.03837
#> 44   0.7162750854 -0.36205399 -0.128991550  0.90747372 1.03837
#> 45   0.6606791397 -0.18572255 -0.172429816  0.87609122 1.03837
#> 46   0.6036817213 -0.02254116 -0.199514066  0.84493996 1.03837
#> 47   0.5452830892  0.12749026 -0.210244040  0.81401988 1.03837
#> 48   0.4854829355  0.26437140 -0.204619979  0.78333110 1.03837
#> 49   0.4242807305  0.38810223 -0.182642845  0.75287375 1.03837
#> 50   0.3625692052  0.49729206 -0.145206397  0.72290629 1.03837
#> 51   0.3045901908  0.58532475 -0.096566917  0.69465929 1.03837
#> 52   0.2508932573  0.65134303 -0.037275536  0.66829214 1.03837
#> 53   0.2014782602  0.69534653  0.032666381  0.64380505 1.03837
#> 54   0.1563456165  0.71733582  0.113259905  0.62119778 1.03837
#> 55   0.1154952223  0.71731083  0.204506430  0.60047023 1.03837
#> 56   0.0789266380  0.69527123  0.306402792  0.58162279 1.03837
#> 57   0.0466402666  0.65121713  0.418951330  0.56465517 1.03837
#> 58   0.0193029875  0.58649490  0.538631079  0.54963921 1.03837
#> 59  -0.0009692641  0.50537678  0.654267610  0.53680302 1.03837
#> 60  -0.0140191103  0.40817940  0.765035100  0.52616325 1.03837
#> 61  -0.0198472255  0.29490253  0.870930936  0.51772020 1.03837
#> 62  -0.0184529342  0.16554646  0.971957589  0.51147357 1.03837
#> 63  -0.0098367392  0.02011104  1.068113692  0.50742356 1.03837
#> 64   0.0060014815 -0.14140364  1.159399370  0.50557009 1.03837
#> 65   0.0290604629 -0.31898617  1.245808372  0.50591216 1.03837
#> 66   0.0582382803 -0.50316225  1.322245823  0.50758659 1.03837
#> 67   0.0914403331 -0.67710680  1.380208988  0.50902516 1.03837
#> 68   0.1284657497 -0.83929591  1.418974995  0.51008269 1.03837
#> 69   0.1693146628 -0.98972945  1.438543843  0.51075918 1.03837
#> 70   0.2139864297 -1.12840786  1.438914731  0.51105482 1.03837
#> 71   0.2615124906 -1.25270101  1.421418144  0.51055728 1.03837
#> 72   0.3090699063 -1.35494956  1.389927442  0.50806622 1.03837
#> 73   0.3564780915 -1.43466385  1.344690183  0.50350493 1.03837
#> 74   0.4037372137 -1.49184385  1.285706076  0.49687338 1.03837
#> 75   0.4508472073 -1.52648942  1.212976090  0.48817149 1.03837
#> 76   0.4978080403 -1.53860080  1.126499404  0.47739937 1.03837
#> 77   0.5446196338 -1.52817789  1.026276297  0.46455700 1.03837
#> 78   0.5912822310 -1.49522067  0.912306480  0.44964436 1.03837
#> 79   0.6377956174 -1.43972919  0.784590519  0.43266144 1.03837
#> 80   0.6841600231 -1.36170329  0.643128187  0.41360820 1.03837
#> 81   0.7303751010 -1.26114318  0.487919022  0.39248475 1.03837
#> 82   0.7764412024 -1.13804872  0.318963702  0.36929100 1.03837
#> 83   0.8223580461 -0.99242009  0.136261478  0.34402704 1.03837
#> 84   0.8681259028 -0.82425706 -0.060186697  0.31669273 1.03837
#> 85   0.9129760362 -0.63614828 -0.265684277  0.28733228 1.03837
#> 86   0.9541094551 -0.43803272 -0.462971339  0.25616156 1.03837
#> 87   0.9890626279 -0.24067617 -0.636334761  0.22359376 1.03837
#> 88   1.0161022165 -0.05352714 -0.774386554  0.19013534 1.03837
#> 89   1.0341080667  0.11466278 -0.869506034  0.15643015 1.03837
#> 90   1.0428340269  0.26190200 -0.920011624  0.12262784 1.03837
#> 91   1.0422800731  0.38819067 -0.925902564  0.08872834 1.03837
#> 92   1.0324462596  0.49352871 -0.887179477  0.05473169 1.03837
#> 93   1.0133324452  0.57791615 -0.803842364  0.02063791 1.03837
#> 94   0.9849387708  0.64135297 -0.675891153 -0.01355302 1.03837
#> 95   0.9472650647  0.68383903 -0.503325628 -0.04784107 1.03837
#> 96   0.9003114768  0.70537477 -0.286146023 -0.08222632 1.03837
#> 97   0.8440780624  0.70595979 -0.024352038 -0.11670872 1.03837
#> 98   0.7785646905  0.68559406  0.282056122 -0.15128824 1.03837
#> 99   0.7038591673  0.64432354  0.632518825 -0.18593823 1.03837
#> 100  0.6246554668  0.58473615  0.997248382 -0.21925759 1.03837
#> 101  0.5464780525  0.51058058  1.341857699 -0.24972883 1.03837
#> 102  0.4717635238  0.42402710  1.651686153 -0.27678012 1.03837
#> 103  0.4005217549  0.32508433  1.926674778 -0.30040913 1.03837
#> 104  0.3342759525  0.21646203  2.159920844 -0.32057200 1.03837
#> 105  0.2754117642  0.10240508  2.340612190 -0.33720005 1.03837
#> 106  0.2241258262 -0.01655875  2.468281889 -0.35033643 1.03837
#> 107  0.1825777640 -0.13429678  2.538597093 -0.36054748 1.03837
#> 108  0.1514272174 -0.24893617  2.550234399 -0.36800614 1.03837
#> 109  0.1306741046 -0.36047659  2.503193621 -0.37271242 1.03837
#> 110  0.1203185229 -0.46891845  2.397475226 -0.37466632 1.03837
#> 111  0.1203602909 -0.57426159  2.233078588 -0.37386778 1.03837
#> 112  0.1307997322 -0.67650587  2.010004070 -0.37031690 1.03837
#> 113  0.1516365208 -0.77565154  1.728251698 -0.36401359 1.03837
#> 114  0.1828385512 -0.87158458  1.388425582 -0.35500762 1.03837
#> 115  0.2232855259 -0.96035115  1.011541949 -0.34502737 1.03837
#> 116  0.2724234832 -1.03999653  0.607989262 -0.33492723 1.03837
#> 117  0.3296399556 -1.10979230  0.184951559 -0.32520350 1.03837
#> 118  0.3928269474 -1.16723249 -0.232849361 -0.31756390 1.03837
#> 119  0.4618071964 -1.21210637 -0.643334301 -0.31215206 1.03837
#> 120  0.5356968696 -1.24450953 -1.040426081 -0.30926438 1.03837
#> 121  0.6090214902 -1.26515910 -1.386986822 -0.31068804 1.03837
#> 122  0.6806696404 -1.27422298 -1.675574113 -0.31677674 1.03837
#> 123  0.7506413163 -1.27170120 -1.906188200 -0.32753046 1.03837
#> 124  0.8189365006 -1.25759358 -2.078828740 -0.34294926 1.03837
#> 125  0.8855550537 -1.23189975 -2.193496185 -0.36303315 1.03837
#> 126  0.9504973467 -1.19462091 -2.250189911 -0.38778202 1.03837
#> 127  1.0137631234 -1.14575615 -2.248910285 -0.41719595 1.03837
#> 128  1.0753522411 -1.08530531 -2.189657732 -0.45127496 1.03837
#> 129  1.1352650492 -1.01326935 -2.072431218 -0.49001894 1.03837
#> 130  1.1935013350 -0.92964720 -1.897231506 -0.53342805 1.03837
#> 131  1.2500610425 -0.83443933 -1.664058570 -0.58150217 1.03837
#> 132  1.3047636938 -0.73329952 -1.399041676 -0.63182791 1.03837
#> 133  1.3573407842 -0.63463094 -1.141013942 -0.68081833 1.03837
#> 134  1.4077897265 -0.53851153 -0.890339987 -0.72843991 1.03837
#> 135  1.4561105488 -0.44494173 -0.647019752 -0.77469258 1.03837
#> 136  1.5023031957 -0.35392169 -0.411053027 -0.81957634 1.03837
#> 137  1.5463677805 -0.26545116 -0.182439946 -0.86309122 1.03837
#> 138  1.5883041971 -0.17953013  0.038819308 -0.90523719 1.03837
#> 139  1.6281124082 -0.09615866  0.252724925 -0.94601425 1.03837
#> 140  1.6657926771 -0.01533701  0.459277162 -0.98542242 1.03837
#> 141  1.7013447540  0.06293539  0.658475305 -1.02346171 1.03837
#> 142  1.7347686587  0.13865806  0.850320385 -1.06013210 1.03837
#> 143  1.7660645102  0.21183103  1.034811375 -1.09543357 1.03837
#> 144  1.7952321429  0.28245470  1.211948685 -1.12936617 1.03837
#> 145  1.8222716669  0.35052846  1.381732546 -1.16192982 1.03837
#> 146  1.8471832042  0.41605277  1.544162919 -1.19312462 1.03837
#> 147  1.8699664972  0.47902745  1.699239446 -1.22295050 1.03837
#> 148  1.8906216970  0.53945245  1.846962380 -1.25140746 1.03837
#> 149  1.9091487291  0.59732817  1.987331415 -1.27849556 1.03837
#> 150  1.9255476436  0.65265426  2.120346990 -1.30421477 1.03837
#> 151  1.9398185547  0.70543028  2.246009069 -1.32856502 1.03837
#> 152  1.9519612771  0.75565719  2.364317169 -1.35154641 1.03837
#> 153  1.9619758973  0.80333422  2.475272134 -1.37315890 1.03837
#> 154  1.9698621978  0.84846214  2.578873002 -1.39340254 1.03837
#> 155  1.9756206293  0.89103985  2.675120518 -1.41227718 1.03837
#> 156  1.9792507766  0.93106865  2.764013967 -1.42978302 1.03837
#> 157  1.9807529528  0.96854718  2.845554468 -1.44591989 1.03837
#> 158  1.9801268358  1.00347683  2.919740454 -1.46068792 1.03837
#> 159  1.9773727300  1.03585656  2.986573437 -1.47408704 1.03837
#> 160  1.9724903644  1.06568655  3.046052472 -1.48611721 1.03837
#> 161  1.9654800076  1.09296730  3.098178061 -1.49677855 1.03837
#> 162  1.9563415132  1.11769810  3.142949685 -1.50607094 1.03837
#> 163  1.9450747457  1.13987963  3.180367786 -1.51399446 1.03837
#> 164  1.9316799627  1.15951134  3.210432367 -1.52054906 1.03837
#> 165  1.9161571396  1.17659362  3.233143482 -1.52573479 1.03837
#> 166  1.8985060768  1.19112622  3.248500275 -1.52955158 1.03837
#> 167  1.8787270983  1.20310891  3.256504408 -1.53199947 1.03837
#> 168  1.8568198401  1.21254232  3.257154524 -1.53307850 1.03837
#> 169  1.8327843643  1.21942668  3.250450171 -1.53278864 1.03837
#> 170  1.8066207708  1.22376078  3.236393213 -1.53112985 1.03837
#> 171  1.7783292128  1.22554554  3.214982187 -1.52810220 1.03837
#> 172  1.7479094461  1.22478049  3.186217641 -1.52370560 1.03837
#> 173  1.7153614351  1.22146584  3.150099146 -1.51794007 1.03837
#> 174  1.6806855996  1.21560202  3.106627420 -1.51080576 1.03837
#> 175  1.6438813488  1.20718824  3.055802047 -1.50230249 1.03837
#> 176  1.6049490181  1.19622530  2.997622310 -1.49243029 1.03837
#> 177  1.5638887763  1.18271224  2.932089342 -1.48118919 1.03837
#> 178  1.5207004544  1.16664931  2.859203483 -1.46857920 1.03837
#> 179  1.4753836730  1.14803712  2.778963446 -1.45460029 1.03837
#> 180  1.4279388916  1.12687613  2.691369381 -1.43925261 1.03837
#> 181  1.3783661368  1.10316471  2.596421822 -1.42253589 1.03837
#> 182  1.3266648935  1.07690449  2.494120896 -1.40445041 1.03837
#> 183  1.2728359611  1.04809363  2.384466105 -1.38499588 1.03837
#> 184  1.2168787000  1.01673389  2.267457693 -1.36417254 1.03837
#> 185  1.1587931059  0.98282479  2.143095488 -1.34198030 1.03837
#> 186  1.0985798671  0.94636501  2.011380684 -1.31841912 1.03837
#> 187  1.0362380998  0.90735739  1.872310009 -1.29348914 1.03837
#> 188  0.9717681904  0.86579900  1.725887526 -1.26719018 1.03837
#> 189  0.9051706673  0.82169032  1.572111357 -1.23952224 1.03837
#> 190  0.8364445225  0.77503340  1.410981190 -1.21048555 1.03837
#> 191  0.7655903467  0.72582665  1.242497305 -1.18007996 1.03837
#> 192  0.6962513499  0.67691047  1.075904885 -1.15000153 1.03837
#> 193  0.6346787895  0.63316132  0.927072508 -1.12316164 1.03837
#> 194  0.5809802817  0.59466270  0.796271631 -1.09961014 1.03837
#> 195  0.5351557819  0.56141404  0.683501167 -1.07934680 1.03837
#> 196  0.4972052592  0.53341664  0.588764113 -1.06237213 1.03837
#> 197  0.4674544618  0.51123296  0.513815945 -1.04896696 1.03837
#> 198  0.4462036607  0.49538864  0.460281111 -1.03939199 1.03837
#> 199  0.4334533710  0.48588124  0.428160933 -1.03364693 1.03837
#> 200  0.4292032196  0.48271257  0.417453343 -1.03173191 1.03837
#> 
#> 
#> 
#> $terms
#> [1] "s(wt)"
#> 
#> $model
#> 
#> Family: gaussian 
#> Link function: identity 
#> 
#> Formula:
#> mpg ~ s(wt)
#> 
#> Estimated degrees of freedom:
#> 2.14  total = 3.14 
#> 
#> GCV score: 7.887675     
#> 
#> $eps
#> [1] 1e-07
#> 
#> $eval
#>           wt
#> 1   1.513000
#> 2   1.532653
#> 3   1.552307
#> 4   1.571960
#> 5   1.591613
#> 6   1.611266
#> 7   1.630920
#> 8   1.650573
#> 9   1.670226
#> 10  1.689879
#> 11  1.709533
#> 12  1.729186
#> 13  1.748839
#> 14  1.768492
#> 15  1.788146
#> 16  1.807799
#> 17  1.827452
#> 18  1.847106
#> 19  1.866759
#> 20  1.886412
#> 21  1.906065
#> 22  1.925719
#> 23  1.945372
#> 24  1.965025
#> 25  1.984678
#> 26  2.004332
#> 27  2.023985
#> 28  2.043638
#> 29  2.063291
#> 30  2.082945
#> 31  2.102598
#> 32  2.122251
#> 33  2.141905
#> 34  2.161558
#> 35  2.181211
#> 36  2.200864
#> 37  2.220518
#> 38  2.240171
#> 39  2.259824
#> 40  2.279477
#> 41  2.299131
#> 42  2.318784
#> 43  2.338437
#> 44  2.358090
#> 45  2.377744
#> 46  2.397397
#> 47  2.417050
#> 48  2.436704
#> 49  2.456357
#> 50  2.476010
#> 51  2.495663
#> 52  2.515317
#> 53  2.534970
#> 54  2.554623
#> 55  2.574276
#> 56  2.593930
#> 57  2.613583
#> 58  2.633236
#> 59  2.652889
#> 60  2.672543
#> 61  2.692196
#> 62  2.711849
#> 63  2.731503
#> 64  2.751156
#> 65  2.770809
#> 66  2.790462
#> 67  2.810116
#> 68  2.829769
#> 69  2.849422
#> 70  2.869075
#> 71  2.888729
#> 72  2.908382
#> 73  2.928035
#> 74  2.947688
#> 75  2.967342
#> 76  2.986995
#> 77  3.006648
#> 78  3.026302
#> 79  3.045955
#> 80  3.065608
#> 81  3.085261
#> 82  3.104915
#> 83  3.124568
#> 84  3.144221
#> 85  3.163874
#> 86  3.183528
#> 87  3.203181
#> 88  3.222834
#> 89  3.242487
#> 90  3.262141
#> 91  3.281794
#> 92  3.301447
#> 93  3.321101
#> 94  3.340754
#> 95  3.360407
#> 96  3.380060
#> 97  3.399714
#> 98  3.419367
#> 99  3.439020
#> 100 3.458673
#> 101 3.478327
#> 102 3.497980
#> 103 3.517633
#> 104 3.537286
#> 105 3.556940
#> 106 3.576593
#> 107 3.596246
#> 108 3.615899
#> 109 3.635553
#> 110 3.655206
#> 111 3.674859
#> 112 3.694513
#> 113 3.714166
#> 114 3.733819
#> 115 3.753472
#> 116 3.773126
#> 117 3.792779
#> 118 3.812432
#> 119 3.832085
#> 120 3.851739
#> 121 3.871392
#> 122 3.891045
#> 123 3.910698
#> 124 3.930352
#> 125 3.950005
#> 126 3.969658
#> 127 3.989312
#> 128 4.008965
#> 129 4.028618
#> 130 4.048271
#> 131 4.067925
#> 132 4.087578
#> 133 4.107231
#> 134 4.126884
#> 135 4.146538
#> 136 4.166191
#> 137 4.185844
#> 138 4.205497
#> 139 4.225151
#> 140 4.244804
#> 141 4.264457
#> 142 4.284111
#> 143 4.303764
#> 144 4.323417
#> 145 4.343070
#> 146 4.362724
#> 147 4.382377
#> 148 4.402030
#> 149 4.421683
#> 150 4.441337
#> 151 4.460990
#> 152 4.480643
#> 153 4.500296
#> 154 4.519950
#> 155 4.539603
#> 156 4.559256
#> 157 4.578910
#> 158 4.598563
#> 159 4.618216
#> 160 4.637869
#> 161 4.657523
#> 162 4.677176
#> 163 4.696829
#> 164 4.716482
#> 165 4.736136
#> 166 4.755789
#> 167 4.775442
#> 168 4.795095
#> 169 4.814749
#> 170 4.834402
#> 171 4.854055
#> 172 4.873709
#> 173 4.893362
#> 174 4.913015
#> 175 4.932668
#> 176 4.952322
#> 177 4.971975
#> 178 4.991628
#> 179 5.011281
#> 180 5.030935
#> 181 5.050588
#> 182 5.070241
#> 183 5.089894
#> 184 5.109548
#> 185 5.129201
#> 186 5.148854
#> 187 5.168508
#> 188 5.188161
#> 189 5.207814
#> 190 5.227467
#> 191 5.247121
#> 192 5.266774
#> 193 5.286427
#> 194 5.306080
#> 195 5.325734
#> 196 5.345387
#> 197 5.365040
#> 198 5.384693
#> 199 5.404347
#> 200 5.424000
#> 
#> $unconditional
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "fderiv"